Get Go traces over grpc (#9846)
This lets `pachctl debug trace` download a Go trace. I also added some
"tasks" so you can figure out what pachd is doing when looking at the
trace:


![Capture](https://github.com/pachyderm/pachyderm/assets/2367/0b4f1bee-d423-40f4-ba08-0c349ad7d5ce)


Tracing has minimal overhead when disabled, and it is only enabled
during the Debug.Trace RPC. Only one user can get a trace at a time.

I implemented this as a new RPC for more convenient testing. We don't
know how enabling traces breaks things, so it's not done by default in
the debug dump. But it would be nice to get one in there, in the future,
after we trust it.