Vendor pfsdb@2.8.6 for clusterstate tests (#9874)
This adds `pfsdb` as it existed in v2.8.6 to the clusterstate directory,
so that the v2.8.0 upgrade tests look at the v2.8.0 database with a
v2.8.0 database library. v2.10.x adds new columns to the tables there,
so src/internal/pfsdb as it exists in v2.10.x can't be used anymore.

I deleted the tests when copying things over; the tests don't work
because they migrate up to the current version, and a bunch of internal
APIs they use have since changed. I think it's cleanest to do things
this way; if the 2.8.x pfsdb code needs to change, change it in that
branch and then copy in the library again.