Add metadata to commits (#9872)
This allows users to set metadata on commit objects. `pachctl edit
metadata commit project/repo@branch-or-commit <op> <data>`. Commits are
parsed by pfs.CommitPicker.UnmarshalText such that the project can be
omitted, and it will be filled in by pachctl from the current context.
This is a new mechanism but consistent with existing behavior.

I adjusted pfsdb tests to use cmp.Diff for comparing commits, instead of
a bunch of require.Equals. Now you can see all the problems with your
commit instead of having to run the test over and over again to see the
next problem.

I adjusted cmdutil so that pachctl bash tests can run as subtests.

This branch is currently based on another one that fixed the
clusterstate tests.