Bazelify etc/generate-envoy-config (#9868)
This just cleans up regenerating envoy.json and friends.

We add a rule `copy_to_workspace` that invokes `prototar` to do our
dirty work. `//src/proto:run` is now implemented in terms of this rule,
and `//etc/helm/pachyderm:update_envoy_config` is too.
`copy_to_workspace` automatically generates a test that reminds you to
regenerate files when necessary, and the test can be hermetic and
cacheable if you know the names of the files you're generating at build
time. (For protos, we don't. For envoy configs, we do.)

Future work in this category involves running jsonnetfmt tests and
jsonnet-lint tests on the jsonnet files; this is currently done outside
of bazel. It would also be nice to run the envoy config validation tests
inside bazel. That requires writing something to copy envoy out of a
docker container because they don't have binaries for linux.