Add auth rules to metadata (#9886)
This makes the metadata server aware of auth; the same rules that are
required to update other fields on the object are checked here.
(Metadata doesn't relax or restrict permissions.)

Something I noticed is that anyone can update commits, which we might
have a bug open for. Anyone can also update project descriptions, which
will need to change.

This also makes `testpachd` capable of activating auth.