read CommitInfo before overwriting it with stale data (#9892)
Refresh CommitInfo before overwriting it.  This avoids a rare race:

A commit is created.
Finalization starts.
Metadata on the commit is changed.
The metadata update is overwritten when the PFS master writes the
finalization time / status to the commit.

The metadata tests are fast enough to cause this, and it happened pretty
consistently.

This probably needs to be done in more places; we should always read
before writing when only changing one field.