Add cluster metadata (#9912)
Cluster metadata is stored in a new table with 1 row, enforced with a
CHECK constraint.

I refactored `pfsdb.jsonMap` into `pgjsontypes.StringMap` so that a new
`coredb` package can use it.

I adjusted InspectCluster to return the metadata, and added the `--raw`
and `--output` options to the CLI so that you can inspect metadata
you've added.

I added a new picker for clusters that just selects the current cluster;
since pachd can't talk to other clusters, there is no need to be able to
address them with pickers. But having a picker made more sense than
`cluster: true` or something.

Auth is handled with a new permission `EDIT_CLUSTER_METADATA` that is
given to clusterAdmins. If people want something like a `metadataOwner`
role, we could add it, but I'm not sure if people want to delegate that
responsibility.

---------

Co-authored-by: Robert Uhl <robert.uhl@hpe.com>