Attempt to deflake the consistenthashing tests (#9943)
I can get this to fail when I run 64 copies of the tests simultaneously.
While I was doing that I noticed that the etcd logs were complaining
about some operations being too slow. Since no CPU was being used while
this was going on, I assumed it was a disk problem. Disabling fsync for
etcd made the parallel runs all complete in a max of 30 seconds or
average of 7 seconds, down from 600 seconds and timing out. So I think
this is the magic sauce that was needed. It probably helps other tests,
since pretty much everything eventually depends on testetcd.

While working on fixing this test, I noticed that the Loki tests are
pretty flaky. That is fixed in #9945.

---------

Co-authored-by: Robert Uhl <robert.uhl@hpe.com>