Find and fix more nested transactions (#9942)
This adds a DPANIC log message for WithTx calls that call WithTx, i.e.
nesting database transactions. Nested database transactions are
problematic because they can deadlock; if the first tx takes the last
database connection out of the pool, the second WithTx will block until
something puts it back in the pool. This trends towards infinity; as
evidenced by CORE-2236.

As part of this change, I made the `pctx.TestContext` logger actually
panic during DPANIC logs, and fixed the places that panicked. (It was
only nested database transactions.) So now `log.Dpanic` will always fail
tests, which has been a goal of mine for a while. Panics do not happen
in production; you only get a stack trace and a log message there. If a
bug sneaks in, it's not a big deal.

To ensure that the WithTx check is actually being used, I have adjusted
the other transaction-starting helpers, transactionenv.WithReadContext
and WithWriteContext to propagate WithTx's context to the callback. That
way, any nesting that happens inside a WithReadContext block will
trigger the panic. I found 0 cases that actually do this.

This change is part of CORE-2246.