Remove nested transactions (#9953)
This removes a couple cases where WithReadContext is run inside WithTx,
which ends up consuming 2 connections from the database connection pool
in a way that can deadlock if there is only 1 connection available in
the pool.

This is sort of a backport of #9942, but without
infrastructure/refactorings to detect nested transactions. This should
be sufficient to eliminate any existing problems, as long as we don't
add more code directly to this branch.

Part of CORE-2246.