Add a parser for the metadata CLI (#9908)
This replaces an inflexible parser that's 1 line of code with a more
complicated one that can parse any valid metadata. I changed the
"replace" metadata op to `replace` in the CLI. `set` is short, but I can
never remember it.

The grammar looks like this:


![kvs](https://github.com/pachyderm/pachyderm/assets/2367/d3a2bfdc-2c1b-41ae-bbdc-85051d2d819a)

`add` and `edit` take a `KV`.  `replace` takes a `KVs`.

Double-quoted strings are the same as Go's, except that `\U12345678`
syntax isn't supported because the spec doesn't explain how it works and
I don't know how it works.

Single-quoted strings are very simple; they can't have any escapes in
them.

Our variant of JSON is more liberal than the official JSON spec out of
convenience. You can now have vertical linefeeds and `\x??` in your
JSON. You can also have a trailing comma. Finally, you can probably
write `{"key":"value""key2":"value2"}`.

I have fuzz tested the parser for about 12 hours x 32 cores, because it
will panic when the parser selects a path that causes the lexer to not
advance. These are mostly caused by `@@*` ("pick 0 or more of this big
subobject") directives; that is why an entirely empty KVs is handled by
checking that `strings.TrimSpaces() == ""`. Allowing empty KVs in the
grammar is hard to get right without lookahead (which makes error
messages bad); lookahead is especially bad for this grammar because how
many tokens you need to look ahead depends on the exact content of
double-quoted strings (because `(escape | double quoted chars)*` is the
content of the string and different strings have different numbers of
tokens). So we avoid lookahead entirely and deal with requring some
valid data to parse. The JSON parser is also not the ideal
implementation which is why quirks like omitting commas works. Frankly,
JSON made its grammar harder to implement than it needed to be to make
it more restrictive than it needed to be. With all that in mind,
although we allow a bunch of weird stuff, we don't reject or misparse
anything valid, so I'm happy.