Improve database and RPC logging  (#9948)
This cleans up some logging. WithTx reports its retry attempt number as
a field instead of being in the logger name, and associates with
postgres pid and transaction id with any logs that are generated in the
transaction. This is most useful with query logging enabled. WithTx is
also a span now, so we can track the duration and success/failure status
of individual transactions easily.

The GRPC logging interceptor has been split into two parts. The first
part runs early and creates a context containing the rpc name, request
id, rpc peer address, etc. This means that any logs that other
interceptors generate, notably auth, are associated with the RPC. The
second part announces the start and end of the RPC, and runs after auth,
so the authenticated username (if any) is always logged here. (The
username is also logged for any other logs generated by the RPC
handler.)

This fixes an issue where database operations done as part of
authorizing the RPC didn't show up with the request ID, etc. (Actually,
that wouldn't have worked until #9947 anyway, so this is more of a fast
follow on top of that.)

It should now be possible to associate postgres logs and x-request-id,
so we can have an end to end view of the system.

Part of CORE-2248