Add context information to logs (#10016)
This adds a `contextExpiresIn` field to every log message logged with a
context that has a deadline. It also adds `contextDone: true` and
`contextError: context.Cause(ctx)` to any logs after a context is done.

It looks something like this; the context expires between `before` and
`after`:

```
=== RUN   TestDeadlineSpan
    logger.go:130: 2024-05-02T20:42:51.104Z     DEBUG   Test    log/span_test.go:362    Test: span start        {"contextExpiresIn": "4.934999ms"}
    logger.go:130: 2024-05-02T20:42:51.104Z     INFO    Test    log/span_test.go:363    before  {"contextExpiresIn": "4.880897ms"}
    logger.go:130: 2024-05-02T20:42:51.109Z     INFO    Test    log/span_test.go:365    after   {"contextExpiresIn": "-225.127µs", "contextDone": true, "contextError": "context deadline exceeded"}
    logger.go:130: 2024-05-02T20:42:51.109Z     DEBUG   Test    log/span_test.go:366    Test: span finished ok  {"contextExpiresIn": "-317.329µs", "contextDone": true, "contextError": "context deadline exceeded", "spanDuration": "5.139824ms"}
--- PASS: TestDeadlineSpan (0.01s)
```

This adds about 166 nanoseconds to every logging operation. That's about
13% overhead, but is a pretty small number. There is no overhead for
Debug() calls when debug logs are turned off, however. (This
optimization can be applied to all the log levels, but realistically
Info is never turned off, so I didn't bother.)

Because we're adding the fields at the last possible moment, I had to
manually find everything that calls `(*zap.Logger).<Log Function>`. I
think I did that successfully.

I found this very useful for debugging flakiness in the compaction test.
Eventually after enabling query logging (and a version of this PR) in
the test, I found that an SQL query was exceeding the deadline by a few
microseconds; that was what made the test flaky. Since it was useful
once, I expect it to be useful again in the future.

I also made it possible to set a deadline in pachctl, so it's easy to
see this feature in action. (Someday we'll want an interceptor to set
deadlines on "fast" calls. Not doing that from console led to an outage
when we were leaking database connections; if we restricted RPCs to some
reasonable time bound, the system would have eventually fixed itself.)