Unignore the pulumi code and add a test for it (#10017)
There was an issue with a dependency of Pulumi that prevented us from
building with Bazel. That is fixed, or at least worked around cleanly:
https://github.com/bazelbuild/rules_go/issues/3799#issuecomment-1880035642

With the workarounds, this now builds OK, so I've unignored it and added
a BUILD file.

There was also an issue where it generates invalid helm values. I fixed
that bug and added a test.