[Backport] Fixes for pipeline extension request handling (#10031) (#10046)
Fixes the way that we handle bad input specs and empty requirements
files. Before, an empty requirements file would cause the pipeline
creation to fail because it was defaulting to the home directory of the
container, since the requirements file was an empty string and not
`None`. Now, it correctly defaults to no requirements file. Bad input
specs before would throw 500 responses, which is incorrect given it is
due to user error.

[INT-1261](https://pachyderm.atlassian.net/browse/INT-1266)

[INT-1261]:

https://pachyderm.atlassian.net/browse/INT-1261?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

[INT-1261]:
https://pachyderm.atlassian.net/browse/INT-1261?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[INT-1261]:
https://pachyderm.atlassian.net/browse/INT-1261?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Kevin Yang <kevin.yang@hpe.com>