avoid logging span start with duplicate fields (#10051)
Prior to this change, the span start logged all fields twice. The fields
make their way into the CheckedEntry that we Write because they were
added on the line above. We can't detect this, because our test
capturing logger uses the Go JSON parser which doesn't have a way to
detect duplicate fields, so we simply never noticed :(

TestDeadlineSpan is one of those tests that you have to manually review
the output of, but I made it so that this mistake is visible and fixed
it.

The original bug didn't make into 2.10.x, so there is no hurry on the
review.