[PFS-247] CDR initial implementation (#10097)
This PR contains the initial implementation of Common Data Refs (CDRs).
CDRs were originally proposed in the following design document:
https://www.notion.so/Common-Data-Refs-3a8e5bf05ecc40bcaf174813581efb92.
A followup design document was created that built on this initial
proposal:
https://www.notion.so/Common-Data-Refs-e22b67d1833340de8da0a7de4e32af26.
This PR implements the first design rather than the second design for
two reasons: a python resolver already exists for the first design and
we would like some experience working with CDRs before committing to the
additional structure in the second design. So, the hope with this
implementation is that it should be easy to get up and running with CDRs
using the existing resolvers, and some time experimenting with CDRs
should inform whether we should introduce the additional structure in
the second design. Also, there are a few known opportunities for
performance improvement, but they should generally only be noticeable
when working with smaller files.