[PFS-68] Normalize Commit Totals (#10037)
This PR migrates the total fileset storage from the `pfs.commit_totals`
table to the `pfs.commits` table. It includes migration code and an
update to the commit store API so that `commitWithID` is used instead of
re-resolving commits multiple times.

A later PR will be added to normalize the pfs.commit_diffs table.