fix broken alpha release (#10108)
`goreleaser` was naming the archive `pachyderm_*` not `pachd_*` like I
wrongly assumed, also removed `--snapshot` as that would add a funky
`snapshot` string to the archive name that we don't want. we have the
`--skip-publish` and `--skip-validate` so we don't need that `--snapshot
`flag will have same effect without there being a `snapshot` string on
the archive.