[PFS-229] Implement ancestor-of and branch-root commit picker parsing (#10099)
This PR adds parsing of "^" and "." (ancestry reference)

"^" refers to git's "^" syntax. It supports the following in this PR:
1. images@master^^
2. images@master^2

Git does not have "." syntax. It supports the following in this PR:
1. images@master.13
2. And it can be used with "^": images@master.13^1 = images@master.12


Parser checks following circumstances:
1. ancestry reference should only contain "^" "." and digits
2. If there is branch root ".", it should be the first operator; It also
should be the only "."; so images@master^1.1 and images@master.1.1 are
not allowed.
3. There should be a number right after "."; images@master.^ is not
allowed.
4. If branch root and ancestor of are used together, the offset of
branch root should >= offset of ancestor of