Djanicek/infeng 734/tests to datadog (#10132)
We have DataDog now in a non-trial capacity. With a little
instrumentation, we can effectively replace all the grafana stuff I did
before with DataDogs version which happens to be basically the same
thing, but with all the data from all the ai-at-scale repos.

This PR hooks up the integration, deploy and bazel tests. to datadog so
now we can get individual test statistics which should help with over
time results and flaky test detection. Here's the test from this branch!
Let me know if you need access.
[Search](https://app.datadoghq.com/ci/test-runs?query=test_level%3Atest%20%40git.branch%3A%22djanicek%2Finfeng-734%2Ftests-to-datadog%22&agg_m=count&agg_m_source=base&agg_t=count&fromUser=false&index=citest&start=1718048675313&end=1718653475313&paused=false)

This change should let us deprecate the test results that I was
collecting in grafana. This collects all that data plus some more.