Set the worker log level to DEBUG, filter to INFO logs when returning logs (#10145)
Worker debug logs contain metrics and other useful information. So we
enable debug logs by default now. GetLogs will only return INFO or
above, if no log level filter is specified.

The logs.LogLevel proto enum is adjusted to explicitly request a log
level, rather than defaulting to debug. This lets us distinguish between
"I want debug logs" and "I want the default".

This also adds log level filtering to GetLogs and `pachctl logs2
--level=...`.

While writing this, I noticed we fail to produce a log object if there
are duplicate keys in the log line. That is allowed by zap and Go's
default JSON decoder, but not by protojson. So we do a normal JSON parse
to extract "severity" for filtering, avoiding the cases where the object
can't be produced. I also adjusted promutil to not use `method` for the
`http_method`, as this is the most frequently duplicated log object key.