starlark: avoid the need to read $PWD (#10153)
It seems the pachdev containers don't run pachd in a readable $PWD. This
is fine, we shouldn't be looking at the filesystem during RunProgram
anyway.

The reason that we do filepath.Abs is so `load()` works relative to the
location of the script file. Since RunProgram isn't running programs
from the filesystem, there isn't really anything that makes sense for
`load` to use as the script location.