[Python SDK] Implement local fileset caching and assembling (#10162)
Implements the ability to create, modify, and utilize a local cache of
PFS chunks when assembling a PFS fileset locally.
The cache stores the compressed and encrypted PFS chunks as files within
a cache directory. The files names for these chunks are the hex-strings
of their hash values prefixed with the name of the hashing algorithm.

* Adds a `PACH_CDR_CACHE_LOCATION` environment variable that users can
configure.
* Extends the Storage API with `fetch_chunks` and `assemble_fileset`
methods.
* Modifies the `CdrResolver` such that it can be configured to use and
modify a local chunk cache.