Implement copy file in fileset API and input fileset creation (#10176)
This PR implements copy file functionality that is made available
through the fileset API. This PR also adds code that creates a fileset
before the processing of a datumset that contains the input files for
the datums in the datumset. This fileset is exposed to user code through
the `FILESET_ID` environment variable and has the same format as the
input files in the meta commit (i.e. `/pfs/<datum-id>/<path>`). This
functionality enables the usage of CDRs within PPS pipelines.

The copy file functionality is implemented through an additional message
type supported by the `storage.CreateFileset` RPC. This functionality is
analogous to the copy file functionality available through PFS, but it
operates on filesets rather than commits. The performance
characteristics of this functionality should be pretty good since it
will only copy references for files > 1MB and it takes advantage of
multiple storage layer caches (i.e. fileset id -> fileset metadata and
file path -> file index).

To create the filesets with the input files, we use the existing fileset
for the datumset that contains the list of datums in the datumset (i.e.
the meta files) and the new copy file functionality. This functionality
is implemented in `datum.CreateInputFileset`, which will iterate through
the datums and send a `CopyFile` message for each file encountered. This
functionality should be pretty straightforward, but it is worth noting
that we use the `pfssync.CacheClient` to maintain a cache of commit ->
fileset id mappings. This is already used in the standard download path,
but we need it here as well to ensure good performance since we don't
want to be resolving commit -> fileset mappings for every file.