Handle lines from misconfigured Loki (#10178)
This adds support for transparently handling CRI-ingested-as-Docker and
Docker-ingested-as-CRI log lines, a common Loki misconfiguration. It
also removes the multiple JSON parses in the log server; we parse once
(or twice if JSON inside Docker JSON) and use that parsed data to
populate the severity, the log object, and the PPS log message.

In the case where an application logs a CRI or Docker -format line ...
we will do the wrong thing. But since we are producing the logs
ourselves, we can simply not do that. And we don't.