A few tweaks for `pachctl logs2` (#10217)
This adjusts a couple of quality-of-life issues:

1. Time flags that are unset are not sent to the server. This changes
the default query from "logs starting a month ago to now" to "logs from
the beginning of time", which is something that TimeRangeIterator knows
how to handle efficiently. This means that no matter how old your
pipeline is, you'll always see logs with `pachctl logs2 --pipeline=edges
--user` or something like that.
2. Time flags can be durations instead of absolute times. To get pachd
logs from the last minute, `pachctl logs2 --from=1m`.
3. With --raw, all logs are printed as their object now, so that Loki metadata makes
it into the printed output. An object is now always synthesized when the
loki line doesn't parse as JSON. This lets you always pipe the output
into jlog or similar for further filtering / querying.
4. `logs2` always prints a warning about an invalid log level; consider
"unset" to be valid and don't warn on the default configuration ;)