Avoid paging hints too close to the current time (#10218)
Given a time interval like [x, now], this avoids returning "now" as the
start of the next batch. That's because the absence of logs doesn't mean
that there aren't logs, just that Loki hasn't seen them yet. Near the
current time, it probably means the logs haven't made it to Loki yet. So
rather than skipping logs that haven't been ingested, when we're near
the current time and haven't seen any logs, we calculate the next paging
hint based on some hard-coded ingestion lag instead. If we have seen a
timestamp, none of this matters; we return a nanosecond after the last
log as the next log.