[PFS-261] Forget RPC Additional Guardrails (#10200)
In this PR we want to check two things:
1. Branch head cannot be moved to a forgotten commit
2. Branch head cannot be forgotten


The TestForgetRPC test is also modified for two reasons:
1. We were checking if the only chunk in the chunk object table is
cleaned. But now, to make a commit forgettable, that commit has to have
a child (so that it's not the branch head). That child's total file set
is also pointing to the chunk, so the chunk will not be cleaned.
2. I synced with Brendon. He thinks we should not test the object
storage level. Testing the ReadFileSet function is sufficient.