[Jupyter] Don't show "invalid address" when cluster is valid (#10225)
This fixes a bug where, when a user starts up the the extension with no
config information and sets the cluster address for the first time, the
"invalid address" error message shows briefly before the extension
redirects, even when the address is valid. This was occurring because
after calling `PUT /config` with the new cluster address, the error
handling used the stale cluster state instead of the response from the
server. This change uses the response from the server to determine if an
error message should be displayed to the user.