[PFS-171] Replace UpdateCommit() (#10231)
The goal of PFS-171 is to update the pfsdb.Commit struct to resolve
resources at the top of RPCs and then use them once throughout. In order
to facilitate that, we've split UpdateCommit() into functions that are
smaller in scope and update a subset of the fields in a commit. Most of
these updates happen to the branch. These are slated to be removed
eventually so having the in a separate function is handy. Next, there is
a function to update the metadata and a function to set some fields once
a commit has been finalized.