[INT-1297] Large repo infinite scroll testing documentation (#10247)
I wrote down some further notes about how to test the infinite
scrolling. The actual implementation is documented already in
MountDrive.tsx. These are just some of the behaviors I am looking for
when I manually test this feature.

It could be possible to write an automated test for this, but since it
can take a while to actually create the repo we might need to use faked
responses from the server. This should still meaningfully test the
frontend.