[PFS-171] Update onUpsert Interface, Refactor StartCommit (#10253)
1. onUpsert now passes through the pfsdb.Commit
2. There's some weird branch and commit logic that we have to do as a
result of our data model that I figured I'd extract into a single
function to make deprecation later easier to do.

The next PR will be an update that actually creates an internal commit
model that we can use. This will be handiest in functions like inspect
commit which have to re-resolve the commit handles for provenance and
such.