[MLDM-24] Include console backend and frontend tests in `bazel test` (#10274)
Creates the `//console/backend:backend_tests` and
`//console/frontend:frontend_tests` bazel test targets. This means that
they will be included when running `bazel test ...`.

Getting the `frontend` tests included was problematic due to the
`canvas` module not having ARM builds published. We decided to manually
exclude the canvas module from the `pnpm` lock file for the frontend
project and exclude the 4 tests which required this dependency.