Add a Gazelle plugin to ensure jsonnet files are linted (#10295)
This adds a new Gazelle plugin that creates `jsonnet_lint_test` rules
for any jsonnet files that appear in the repository. If the file is
deleted, the rule is deleted, of course.

I want to do this for shell scripts next, so we can add shellcheck rules
for each of them and remove the global shellcheck linter.

The new plugin got some that I missed in the previous PR, as well. I
wish I had written this first.