Move all golangci-lint linters to nogo (#10276)
This moves all linters from golangci-lint to nogo. golangci-lint is
slow, and is broken on a version that works with go 1.23 because of some
changes internal to staticcheck. Because I want to use go 1.23 AND have
all the staticcheck tests work, I moved all the linters to nogo.

nogo lints code as it's compiled, so it's cached, and you never have to
manually lint go files. It's also very easy to add our own linters now;
add a package to //src/internal/analyzers, and you can immediately start
using it.

Doing this conversion uncovered a bunch of stuff that golangci-lint
hides by default. This PR fixes the issues (mostly errcheck) or disables
the underlying linter; a follow-up will enable more linters.

This is part 1 of MLDM-42.