upgrade docker to resolve a govulncheck warning (#10315)
Upgrading to the very latest version of the docker go client results in
an issue where it depends on an old version of otel, but our etcd client
depends on a newer version, and they are incompatible with each other,
causing the etcd client to be downgraded. I don't want to downgrade
etcd, so I picked a version of docker with the security fix but without
the otel dependency and upgraded that. We also depend on
fsouza/go-dockerclient (a wrapper) and I picked a version that is
compatible with the version of Docker we are switching to.

Using this client requires us to update Docker for our machine jobs. I'm
using Ubuntu 24.04 now. This, of course, broke some Python jobs which
use the host Python, which was also upgraded. Ben fixed the label studio
example and I merged his PR into this one. I fixed the jupyter tests by
explicitly installing a Python version with pyenv. Obviously, both of
these jobs should use Bazel and get Python that way, but they don't yet.

While I was in there, we had some code that attempted to test against
Python 3.8, 3.9, and 3.10. It didn't actually work (the job runs 3
times, but uses the host Python all 3 times), and to make it work
requires a lot of refactoring that didn't go well for me, so I just
deleted it. Ben and I discussed it, and we decided to handle bugs in old
Pythons on a case by case basis.

Finally, there are a bunch of security problems in the go 1.22.5
standard library, so this PR moves us to 1.22.7. 1.23.1 will be trivial
and merged next, but I don't want to do a major version upgrade in this
PR.

This fixes https://pkg.go.dev/vuln/GO-2024-2512

Tracking issue: MLDM-77

---------

Co-authored-by: Ben Bonenfant <benjamin.bonenfant@hpe.com>