Add snapshot database schema (#10358)
This adds the basic snapshot database schema, and a package to put the
business logic in.

I have stubbed out some steps for creating snapshots, just to ensure the
foreign keys work as expected.

I also added a function to dockertestenv to return a database with
migrations already applied. This pattern is cut-n-pasted into a ton of
places. I'll clean that up in the near future.