Let psql handle C-c  (#10366)
Pressing Control-C while psql is running cancels the context that runs
psql, which kills psql with signal 9. This is annoying as Control-C is
used by psql for other things. This PR changes the behavior to ignore
the cancellation behavior on the go side, so that psql receives the
signal and can cancel ongoing queries, etc.