Add a function to restore a snapshot (#10373)
This adds the ability to restore the database state to that contained in
a snapshot. After restoring, the snapshot is still restorable, because
we re-upload the database dump and adjust the snapshot to show the new
ID. This makes it look like snapshots contain themselves, which is a
nice feature ;)

To restore cleanly, we can't rely on queries running against a default
schema (public), it seems, so I updated the migrations code to be
explicit about the schema the queries are in.

`psql` is used to restore the dump, because that's what the `pg_dump`
documentation says to do. I thought about parsing the SQL and sending
one statement at a time, but it seems like something that is likely to
go wrong in weird ways. As a result of this, `psql` is available in
//tools/postgres/psql now, which should be convenient.

We use pg_dump v17 to dump any version of postgres. For various reasons,
these dumps can only be restored onto postgres 17. The snapshotting code
has been adjusted to remove the v17-specific functionality so that we
can take snapshots of v13 and restore onto v13.

Finally, I adjusted the snapshotting code to atomically create the
pachyderm snapshot and the postgres snapshot. This involves sharing the
transaction between Go and pg_dump which is possible (see
pg_export_snapshot). We have to manually dump everything created in that
transaction, though, which we now do. (The start READING the same MVCC
snapshot, but can't see each other's writes.)