Build test and release containers with Bazel (#10397)
This moves the container build out of goreleaser and to Bazel. CI is
still a mess; the container builds get run 3 or 4 times for no good
reason. That is out of scope to fix here. Goreleaser is still used to
build pachd and that stuff; we upload a ton of random stuff to Github
for releases and I didn't bother breaking that out here. Next quarter we
can probably eliminate goreleaser.

I made sure the coverage image actually works; it needed GOOS/GOARCH
passed in. I also removed the default to add --stamp to the builds. This
doesn't do much right now, but will improve cache hits in the future
(for test jobs that need a container image, not stamping means that
merely committing new unrelated files doesn't invalidate those test
results).