Make the CreateSnapshot API actually take a snapshot (#10405)
This moves the recovery package into the snapshot package, and changes
the RPC to call CreateSnapshot instead of just creating an empty
snapshot row. The original package duplicated some stuff (the snapshot
struct), so I've merged that into the one that already exists in the
snapshot package.

I noticed more shared library dependency hell when running this "for
real", so I've changed the container to be built with the full debian
dependency set for postgresql-client-17. This means that things like
perl are in there now, because postgres' wrapper scripts are written in
perl. It's only about 30MB of data, though, so not a big deal.

To test recovery, this PR exposes `pachdev restore --id=<snapshot id>`
to start a restore job. It does not work yet.

Finally, I adjusted the pachctl commands to show more than 100 snapshots
and to show the fileset containing the database dump, so it can be
downloaded with storage.Fileset.ReadFileset.