Re-enable the printf linter (#10415)
I have noticed this rule triggering in the test sharder, causing all
integration tests to fail, which is very annoying. It shouldn't have
been turned off in the first place, but I didn't understand the
multitude of mistakes it was detecting and decided "later". I understand
the mistakes now (someone was very cut-n-paste happy while writing the
require library), and have fixed them. So now you don't have to worry
about the integration tests failing because of a lint error.

MLDM-172