Run Minio on a port that Zscaler hasn't stolen for itself (#10433)
The Zscaler proxy lives on port 9000. This results in the tests (and
test pachd) not working on HPE MacOS devices. Does Docker return an
error when creating a port forward to this already used port? Nope.
Zscaler must be doing something "clever".

After rebasing past this commit, `docker rm -f pach_test_minio` to get
rid of the old minio test container. The new one will be created
successfully regardless, but it's a waste of resources. (This is why the
new one starts at port 9002 and not 9001, btw.)