Remove Enterprise checks and fix tests
fix up tests without enterprise. They all passed! 



Here's what I did. This was Started off Muyang's branch so I could mess
around, but now there are two branches, so I want his input to make sure
I haven't messed anything up in this. I'm also happy to close this and
have changes cherry-picked to a cleaner branch.

- label studio - I just turned auth off. It looks like the tests were
running w/o enterprise before, so this should be the same as it was
before.
- I think python-sdk was getting the the wrong commit in the test . It
wanted to get the input commit for the job, but before it got the
spec_commit which is the commit for the repo that holds the pipeline
spec. IDK how this worked before :sweat_smile: It is failing now because
I think only the internal pfs user has access to the spec commit. I
changed it so input_commit=job_info.details.input.pfs.commit and it is
working even when I don't pass the pipeline/project so I'm reasonably
confident that this is right.
- Console - I made an additional helm chart w/o auth for the community
tests. We may just want to delete it though.

---------

Co-authored-by: MuyangZhangHPE <muyang.zhang@hpe.com>
Co-authored-by: Pierre Lamarche <pierrelamarche5@gmail.com>
Co-authored-by: Alon Cohen <alon.cohen@hpe.com>
Co-authored-by: Fahad Syed <fahad.syed@hpe.com>